[rename🚚] 중간테이블 mapping 폴더에 정리 (#15)
