+++ src/main/java/com/spoteditor/backend/placelog/entity/PlaceLog.java
@@ -2,10 +2,10 @@ package com.spoteditor.backend.placelog.entity;
 
 import com.spoteditor.backend.global.common.BaseEntity;
 import com.spoteditor.backend.place.entity.Address;
-import com.spoteditor.backend.placelogplace.entity.PlaceLogPlace;
+import com.spoteditor.backend.mapping.placelogplace.entity.PlaceLogPlace;
 import com.spoteditor.backend.user.entity.User;
 
-import com.spoteditor.backend.userplacelog.entity.UserPlaceLog;
+import com.spoteditor.backend.mapping.userplacelog.entity.UserPlaceLog;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
